requirements.txt
Django>=4.0
djangorestframework==3.14
psycopg2==2.9.6
PyYAML>=5.1,<5.4
drf-spectacular>=0.15.1
django-extensions
django-cors-headers
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
django-rest-passwordreset>=1.2.1
python-dateutil
uwsgi>=2.0.19,<2.1
flake8==4.0.0
pillow
django-crispy-forms
django-widget-tweaks
django-filter
openpyxl
faker
gunicorn
jinja2
weasyprint
whitenoise
dj-database-url

